Add Robes for Cho and Hermione
This adds Robes for both Cho and Hermione.  There is 1 version for Cho and 3 versions for Hermione.  This also patches many of the outfits for both girls to allow parts of the clothes to poke out above the top fo the robes to look correct and not require you to make different variations of the robes for each item.  It's mostly school outfits for both, but also some other side outfits for Hermione.

I also added the robes to be part of their default outfits.